Towards New Processing Class
Factored process_synthetic_data() out of
test_compare_aurora_vs_archived_emtf.py,
so that the same function process_synthetic_data can be called in test_synthetic_driver,
ie all the synthetic processing uses the same gateway to the pipeline
